Pre prod

+++ tests/Feature/AlumnoControllerTest.php
@@ -13,21 +13,16 @@ use LdapRecord\Models\ActiveDirectory\User;
 
 class AlumnoControllerTest extends TestCase
 {
-    /**
-     * A basic feature test example.
-     *
-     * @return void
-     */
+
 
     use RefreshDatabase;
 
-    public function test_create_user_alumno()
+    public function testCreateUserAlumno()
     {
         $token = token::factory()->create();
-        $padded_number = str_pad(mt_rand(1, 9999999), 1 - strlen('1'), '0', STR_PAD_LEFT);
-        $randomID = "1". $padded_number;
+        $randomID = str_pad(mt_rand(10000000, 99999999), 7);
         $newStudent = [
-            'samaccountname' =>$randomID,
+            'samaccountname' => $randomID,
             'name' => "Thomas",
             'surname' => "Edison",
             'userPrincipalName' => 'tedison@example.com',
@@ -35,52 +30,55 @@ class AlumnoControllerTest extends TestCase
             'grupos' => [],
         ];
 
-        $response = $this->post('api/usuario', $newStudent,[
+        $response = $this->post('api/usuario', $newStudent, [
             'token' => [
                 $token->token,
             ],
         ]);
-        $this->deleteCreatedLDAPUser($newStudent['samaccountname']);
+        /* $this->deleteCreatedLDAPUser($newStudent['samaccountname']); */
         $response->assertStatus(200);
         $response->assertSee($newStudent['userPrincipalName']);
         $response->assertSee($newStudent['ou']);
-    
+
+        $this->assertDatabaseHas('usuarios', [
+            'id' => $newStudent['samaccountname'],
+            'ou' => $newStudent['ou'],
+        ]);
+
     }
     public function deleteCreatedLDAPUser($samaccountname)
     {
-        try{
-            $user = User::find('cn='.$samaccountname.',ou=UsuarioSistema,dc=syntech,dc=intra');
+        $user = User::find('cn=' . $samaccountname . ',ou=Testing,dc=syntech,dc=intra');
+        if (!empty($user)) {
             $user->delete();
-        }catch(\Exception $e){
-            return null;
         }
-        
     }
 
 
-    public function test_list_users_alumno()
+    public function testListUsersAlumno()
     {
         $token = token::factory()->create();
-        $alumno1 =  $this->createNewAlumno();
-   
-        $response = $this->get('api/alumno',[
+        $alumno1 = $this->createNewAlumno();
+
+        $response = $this->get('api/alumno', [
             'token' => [
                 $token->token,
             ],
         ]);
-       
+
         $response->assertStatus(200);
-    
+
         $this->assertEquals($response[0]['id'], $alumno1);
-      
+
     }
 
-    public function test_show_user_alumno(){
+    public function testShowUserAlumno()
+    {
         $token = token::factory()->create();
-      
-        $alumno =  $this->createNewAlumno();
-      
-        $response = $this->get('api/alumno/'.$alumno,[
+
+        $alumno = $this->createNewAlumno();
+
+        $response = $this->get('api/alumno/' . $alumno, [
             'token' => [
                 $token->token,
             ],
@@ -90,10 +88,11 @@ class AlumnoControllerTest extends TestCase
 
     }
 
-    public function test_error_show_user_alumno(){
+    public function testErrorShowUserAlumno()
+    {
         $token = token::factory()->create();
-      
-        $response = $this->get('api/alumno/'."testUser",[
+
+        $response = $this->get('api/alumno/' . "testUser", [
             'token' => [
                 $token->token,
             ],
@@ -101,11 +100,11 @@ class AlumnoControllerTest extends TestCase
         $response->assertStatus(404);
 
     }
-    public function createNewAlumno(){
+    public function createNewAlumno()
+    {
 
-        $padded_number = str_pad(mt_rand(1, 9999999), 1 - strlen('1'), '0', STR_PAD_LEFT);
-        $randomID = "1". $padded_number;
-       
+        $randomID = str_pad(mt_rand(10000000, 99999999), 7);
+
         $user = usuarios::factory()->create([
             'id' => $randomID,
             'ou' => 'Alumno'
@@ -118,7 +117,8 @@ class AlumnoControllerTest extends TestCase
         return $randomID;
     }
 
-    public function test_update_user_alumno(){
+    public function testUpdateUserAlumno()
+    {
         $userID = $this->createNewAlumno();
         $token = token::factory()->create();
         $updatedUser = [
@@ -127,13 +127,13 @@ class AlumnoControllerTest extends TestCase
             'email' => 'jane.doe@example.com',
             'genero' => 'Femenino',
         ];
-    
-        $response = $this->put("api/alumno/".$userID, $updatedUser, [
+
+        $response = $this->put("api/alumno/" . $userID, $updatedUser, [
             'token' => [
                 $token->token,
             ],
         ]);
-    
+
         $response->assertStatus(200);
         $response->assertJson([
             'usuario' => [
@@ -146,7 +146,8 @@ class AlumnoControllerTest extends TestCase
 
     }
 
-    public function test_error_update_user_alumno(){
+    public function testErrorUpdateUserAlumno()
+    {
         $userID = "RandomUser";
         $token = token::factory()->create();
         $updatedUser = [
@@ -155,8 +156,8 @@ class AlumnoControllerTest extends TestCase
             'email' => '2314214',
             'genero' => 'Femenino',
         ];
-    
-        $response = $this->put("api/alumno/".$userID, $updatedUser, [
+
+        $response = $this->put("api/alumno/" . $userID, $updatedUser, [
             'token' => [
                 $token->token,
             ],
@@ -164,13 +165,13 @@ class AlumnoControllerTest extends TestCase
         $response->assertStatus(404);
     }
 
-    //  Route::get('/alumno/{id}/grupos', 'App\Http\Controllers\AlumnoController@gruposNoPertenecenAlumno');
 
-    public function test_list_grupos_no_pertenece_alumno(){
+    public function testListGruposNoPerteneceAlumno()
+    {
         $token = token::factory()->create();
-        $alumno =  $this->createNewAlumno();
+        $alumno = $this->createNewAlumno();
         $grupo = grupos::factory()->create();
-        $response = $this->get('api/alumno/'.$alumno.'/grupos',[
+        $response = $this->get('api/alumno/' . $alumno . '/grupos', [
             'token' => [
                 $token->token,
             ],
@@ -179,11 +180,12 @@ class AlumnoControllerTest extends TestCase
         $this->assertEquals($response[0]['id'], $grupo->id);
     }
 
-    public function test_error_list_grupos_no_pertenece_alumno(){
+    public function testErrorListGruposNoPerteneceAlumno()
+    {
         $token = token::factory()->create();
         $alumno = "randomId";
         $grupo = grupos::factory()->create();
-        $response = $this->get('api/alumno/'.$alumno.'/grupos',[
+        $response = $this->get('api/alumno/' . $alumno . '/grupos', [
             'token' => [
                 $token->token,
             ],
